third_party/blink/web_tests/platform/generic/fast/dom/document-contentType-createDocument-expected.txt

PASS doc.contentType is "text/html"
PASS doc.contentType is "application/xml"
PASS successfullyParsed is true

TEST COMPLETE
